refactor pulumi
