+++ batch_debugging_cli/pulumi_commands.py
@@ -1,11 +1,43 @@
 import os
+from typing_extensions import Annotated
 import typer
 
 from infrastructure.pulumi import PulumiExecution
 from infrastructure.gcp_compute_engine import PulumiGCP
 from infrastructure.pulumi_config import PulumiGCPConfig, PulumiConfig
 
-class PulumiCommandsInterface():
+
+pulumi = typer.Typer()
+
+class PulumiCommands():
+    ##Pulumi
+    @pulumi.command("up")
+    def up(config_file:  Annotated[str, typer.Option(prompt="The location of the YAML file")] ):
+        pulumi_commands = Pulumi(config_file)
+        pulumi_commands.pulumi_up()
+
+    @pulumi.command("destroy")
+    def destroy(config_file:  Annotated[str, typer.Option(prompt="The location of the YAML file")] ):
+        pulumi_commands = Pulumi(config_file)
+        pulumi_commands.pulumi_destroy()
+
+    @pulumi.command("preview")
+    def destroy(config_file:  Annotated[str, typer.Option(prompt="The location of the YAML file")] ):
+        pulumi_commands = Pulumi(config_file)
+        pulumi_commands.pulumi_preview()
+        
+    @pulumi.command("cancel")
+    def cancel(config_file:  Annotated[str, typer.Option(prompt="The location of the YAML file")] ):
+        pulumi_commands = Pulumi(config_file)
+        pulumi_commands.pulumi_cancel()
+    
+    @pulumi.command("destroy-stack")   
+    def destroy_stack(config_file:  Annotated[str, typer.Option(prompt="The location of the YAML file")] ):
+        pulumi_commands = Pulumi(config_file)
+        pulumi_commands.destroy_stack()
+
+
+class PulumiCommands():
     def pulumi_up(self) -> None:
         pass
     
@@ -24,7 +56,7 @@ class PulumiCommandsInterface():
     def select_gcp_type(self):
         pass
 
-class PulumiCommands(PulumiCommandsInterface):
+class Pulumi(PulumiCommands):
     
     def __init__(self, config_file: str) -> None:
         self.config_file = config_file
